# Optimizing SVG conversion build time