After running multiple builds lately for testing, the build time was taking an atrocious amount of time to run. Seeing that the bulk of the build was relying on the SVG conversion shell commands, I thought I would attempt to convert that to a Node script that was a little more efficient. Based on the timings below, we're seeing roughly a 95% decrease in build time, going from over 5 minutes to just under 14 seconds.

**Before**
```
yarn build  50.41s user 15.09s system 20% cpu 5:23.98 total
```

**After**
```
yarn build  7.99s user 2.18s system 73% cpu 13.818 total
```

- Convert bash script to Node script
- If PNG, copy over as-is, don't check
- Loop through SVGs and check if the converted file exists
- Check if SVG source file has recently changed. If it has, then convert it, else just skip it.

There's still some work to be done on efficiencies and a Node warning I'm seeing related to the promises for when SVGs are being converted, but overall it looks promising.

```
(node:2371) MaxListenersExceededWarning: Possible EventEmitter memory leak detected. 11 exit listeners added to [process]. Use emitter.setMaxListeners() to increase limit
(node:2371) MaxListenersExceededWarning: Possible EventEmitter memory leak detected. 11 SIGINT listeners added to [process]. Use emitter.setMaxListeners() to increase limit
```