@kyletaylored 2021-01-20
@AliasIO before I put anymore effort into this, just wanted to get your initial feedback on if this would be valuable to put some more work into, or not really worth it?