@kyletaylored 2021-02-02
Technically, yes? The two main things I needed to look into further were:

1. How to detect a difference in files (this assumes something like a week)
2. Fixing the Node memory error / warnings