@AliasIO 2021-02-02
The warning in not a concern, we can just raise the limit.

For the file diffs we can probably just generate an hash for each file and compare those. I don't think we need the date check.