@AliasIO 2021-05-19
I optimised the script, reusing the Chromium instance helps a lot with memory and crashes. :)